requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.5
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cssutils==2.2.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.5.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
multitasking==0.0.9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
praw==7.1.4
prawcore==1.5.0
premailer==3.7.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psaw==0.0.12
ptyprocess==0.7.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.0
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytrends==4.7.3
pytz==2020.5
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-futures==1.0.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.15.0
stevedore==3.3.0
tabulate==0.8.7
terminado==0.9.2
testpath==0.4.4
toml==0.10.1
tornado==6.1
tqdm==4.55.2
traitlets==5.0.5
TwitterAPI==2.6.9.1
update-checker==0.18.0
urllib3==1.26.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
wrapt==1.12.1
yahooquery==2.2.12
